[Refactor] DPL-080: 코드 리펙토링

+++ domain/src/main/kotlin/com/user/domain/terms/Terms.kt
@@ -2,7 +2,7 @@ package com.user.domain.terms
 
 data class Terms(
     val termsId: Long,
-    val sequence: Long,
+    val sequence: Int,
     val content: String,
     val mandatory: Boolean,
 )
